refactor(cli): split commands into package
